+++ Source/StarGame/Entities/Enemy.h
@@ -237,7 +237,7 @@ private:
 public:
 	Swarm() {}
 	Swarm(int newSwarmersCount, 
-		  int newTime_seconds, int newDamage,
+		  float newTime_seconds, int newDamage,
 		  const BillboardProgDataNoTexture &billboardProgDataNoTexture,
 		  glm::vec4 newInitialColor, glm::vec4 newOnFreezeColor,
 		  glm::vec3 newPosition, glm::vec3 newFrontVector,
